Fixed code smell
git-svn-id: https://satscm.esac.esa.int/esasky/esasky-app/trunk@5320 aa7853a0-da02-4334-bf23-07ffea9f4f01

+++ esasky-cl/src/main/java/esac/archive/esasky/cl/web/client/view/ctrltoolbar/treemap/TreeMapContainer.java
@@ -183,7 +183,7 @@ public class TreeMapContainer extends DialogBox {
 				styleString += ESASkyColors.getColor(i);
 				if(nShown == 1) {
 					styleString += " " + Double.toString(botPosition*100) + "%";
-				}else if(nShown == Math.ceil(high) - Math.floor(low) - 1) {
+				}else if(nShown == (int) Math.ceil(high) - (int) Math.floor(low) - 1) {
 					styleString += " " + Double.toString(100 - topPosition * 100) + "%";
 				}
 				nShown++;
